Time: 0 ms (100%), Space: 26.7 MB (64.24%) - LeetHub

+++ 3427-sum-of-variable-length-subarrays/3427-sum-of-variable-length-subarrays.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int subarraySum(vector<int>& nums) {
-        long long ans=0;
+        int ans=0;
         int n= nums.size();
 
         for(int i=0; i<n; i++){
